+++ pkg/pipeline/input/bin/bin.go
@@ -1,37 +1,100 @@
-package input
+package bin
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/tinyzimmer/go-gst/gst"
 
 	"github.com/livekit/egress/pkg/errors"
-	"github.com/livekit/egress/pkg/pipeline/source"
+	"github.com/livekit/egress/pkg/pipeline/params"
+	"github.com/livekit/protocol/tracer"
 )
 
-type Bin struct {
-	source.Source
+type InputBin struct {
+	bin *gst.Bin
 
-	bin           *gst.Bin
 	audioElements []*gst.Element
 	videoElements []*gst.Element
-
-	multiQueue *gst.Element
-	audioPad   *gst.Pad
-	videoPad   *gst.Pad
-
-	mux *gst.Element
+	multiQueue    *gst.Element
+	mux           *gst.Element
+	audioPad      *gst.Pad
+	videoPad      *gst.Pad
 }
 
-func (b *Bin) Bin() *gst.Bin {
+func New(audioElements, videoElements []*gst.Element) *InputBin {
+	return &InputBin{
+		bin:           gst.NewBin("input"),
+		audioElements: audioElements,
+		videoElements: videoElements,
+	}
+}
+
+func (b *InputBin) Build(ctx context.Context, p *params.Params) error {
+	ctx, span := tracer.Start(ctx, "Input.build")
+	defer span.End()
+
+	var err error
+	if err = b.bin.AddMany(b.audioElements...); err != nil {
+		return err
+	}
+	if err = b.bin.AddMany(b.videoElements...); err != nil {
+		return err
+	}
+
+	// queue
+	b.multiQueue, err = gst.NewElement("multiqueue")
+	if err != nil {
+		return err
+	}
+	if err = b.bin.Add(b.multiQueue); err != nil {
+		return err
+	}
+
+	// mux
+	b.mux, err = BuildMux(p)
+	if err != nil {
+		return err
+	}
+	if b.mux != nil {
+		if err = b.bin.Add(b.mux); err != nil {
+			return err
+		}
+	}
+
+	// HLS has no output bin
+	if p.OutputType == params.OutputTypeHLS {
+		return nil
+	}
+
+	// create ghost pad
+	var ghostPad *gst.GhostPad
+	if b.mux != nil {
+		ghostPad = gst.NewGhostPad("src", b.mux.GetStaticPad("src"))
+	} else if len(b.audioElements) != 0 {
+		b.audioPad = b.multiQueue.GetRequestPad("sink_%u")
+		ghostPad = gst.NewGhostPad("src", b.multiQueue.GetStaticPad("src_0"))
+	} else if len(b.videoElements) != 0 {
+		b.videoPad = b.multiQueue.GetRequestPad("sink_%u")
+		ghostPad = gst.NewGhostPad("src", b.multiQueue.GetStaticPad("src_0"))
+	}
+
+	if !b.bin.AddPad(ghostPad.Pad) {
+		return errors.ErrGhostPadFailed
+	}
+
+	return nil
+}
+
+func (b *InputBin) Bin() *gst.Bin {
 	return b.bin
 }
 
-func (b *Bin) Element() *gst.Element {
+func (b *InputBin) Element() *gst.Element {
 	return b.bin.Element
 }
 
-func (b *Bin) Link() error {
+func (b *InputBin) Link() error {
 	mqPad := 0
 
 	// link audio elements
