+++ pkg/pipeline/input/sdk/source.go
@@ -1,7 +1,6 @@
-package source
+package sdk
 
 import (
-	"context"
 	"strings"
 	"sync"
 	"time"
@@ -9,71 +8,20 @@ import (
 	"github.com/pion/webrtc/v3"
 	"github.com/tinyzimmer/go-gst/gst"
 	"github.com/tinyzimmer/go-gst/gst/app"
-	"go.uber.org/atomic"
 
 	"github.com/livekit/egress/pkg/errors"
 	"github.com/livekit/egress/pkg/pipeline/params"
 	"github.com/livekit/protocol/livekit"
-	"github.com/livekit/protocol/logger"
-	"github.com/livekit/protocol/tracer"
 	lksdk "github.com/livekit/server-sdk-go"
 )
 
-const (
-	AudioAppSource = "audioAppSrc"
-	VideoAppSource = "videoAppSrc"
-
-	subscriptionTimeout = time.Second * 5
-)
-
-type SDKSource struct {
-	room   *lksdk.Room
-	logger logger.Logger
-	active atomic.Int32
-	cs     *clockSync
-
-	// track
-	trackID string
-
-	// track composite
-	audioTrackID string
-	videoTrackID string
-
-	// participant
-	participantIdentity string
-
-	// composite audio source
-	audioSrc     *app.Source
-	audioCodec   webrtc.RTPCodecParameters
-	audioWriter  *appWriter
-	audioPlaying chan struct{}
-
-	// composite video source
-	videoSrc     *app.Source
-	videoCodec   webrtc.RTPCodecParameters
-	videoWriter  *appWriter
-	videoPlaying chan struct{}
-
-	mutedChan    chan bool
-	endRecording chan struct{}
-}
-
-func NewSDKSource(ctx context.Context, p *params.Params) (*SDKSource, error) {
-	ctx, span := tracer.Start(ctx, "SDKSource.New")
-	defer span.End()
-
-	s := &SDKSource{
-		logger:       p.Logger,
-		cs:           &clockSync{},
-		mutedChan:    p.MutedChan,
-		endRecording: make(chan struct{}),
-	}
-
+func (s *SDKInput) joinRoom(p *params.Params) error {
 	cb := lksdk.NewRoomCallback()
+	cb.OnTrackPublished = s.onTrackPublished
 	cb.OnTrackMuted = s.onTrackMuted
 	cb.OnTrackUnmuted = s.onTrackUnmuted
 	cb.OnTrackUnpublished = s.onTrackUnpublished
-	cb.OnDisconnected = s.onComplete
+	cb.OnDisconnected = s.onDisconnected
 
 	var onSubscribeErr error
 	var wg sync.WaitGroup
@@ -162,112 +110,69 @@ func NewSDKSource(ctx context.Context, p *params.Params) (*SDKSource, error) {
 	}
 
 	s.room = lksdk.CreateRoom(cb)
+	s.logger.Debugw("connecting to room")
+	if err := s.room.JoinWithToken(p.LKUrl, p.Token, lksdk.WithAutoSubscribe(false)); err != nil {
+		return err
+	}
 
 	var fileIdentifier string
 	switch p.Info.Request.(type) {
+	case *livekit.EgressInfo_ParticipantComposite:
+		fileIdentifier = p.ParticipantIdentity
+		s.participantIdentity = p.ParticipantIdentity
+		if err := s.subscribeToParticipant(&wg); err != nil {
+			return err
+		}
+
 	case *livekit.EgressInfo_TrackComposite:
 		fileIdentifier = p.Info.RoomName
+		tracks := make(map[string]struct{})
 		if p.AudioEnabled {
 			s.audioTrackID = p.AudioTrackID
-			wg.Add(1)
+			tracks[s.audioTrackID] = struct{}{}
 		}
 		if p.VideoEnabled {
 			s.videoTrackID = p.VideoTrackID
-			wg.Add(1)
+			tracks[s.videoTrackID] = struct{}{}
+		}
+		wg.Add(len(tracks))
+		if err := s.subscribeToTracks(tracks); err != nil {
+			return err
 		}
 
 	case *livekit.EgressInfo_Track:
 		fileIdentifier = p.TrackID
 		s.trackID = p.TrackID
 		wg.Add(1)
-	}
-
-	if err := s.join(ctx, p); err != nil {
-		return nil, err
+		if err := s.subscribeToTracks(map[string]struct{}{s.trackID: {}}); err != nil {
+			return err
+		}
 	}
 
 	wg.Wait()
 	if onSubscribeErr != nil {
-		return nil, onSubscribeErr
+		return onSubscribeErr
 	}
 
 	if p.EgressType == params.EgressTypeFile {
 		if err := p.UpdateOutputTypeFromCodecs(fileIdentifier); err != nil {
 			s.logger.Errorw("could not update file params", err)
-			return nil, err
-		}
-	}
-
-	return s, nil
-}
-
-func (s *SDKSource) join(ctx context.Context, p *params.Params) error {
-	ctx, span := tracer.Start(ctx, "SDKSource.join")
-	defer span.End()
-
-	s.logger.Debugw("connecting to room")
-	if err := s.room.JoinWithToken(p.LKUrl, p.Token, lksdk.WithAutoSubscribe(false)); err != nil {
-		return err
-	}
-
-	expecting := make(map[string]bool)
-	if s.trackID != "" {
-		expecting[s.trackID] = true
-	} else {
-		if s.audioTrackID != "" {
-			expecting[s.audioTrackID] = true
-		}
-		if s.videoTrackID != "" {
-			expecting[s.videoTrackID] = true
-		}
-	}
-
-	deadline := time.Now().Add(subscriptionTimeout)
-	for time.Now().Before(deadline) {
-		err := s.subscribeToExpectedTracks(expecting)
-		if err != nil {
 			return err
 		}
-		if len(expecting) == 0 {
-			break
-		}
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	for trackID := range expecting {
-		return errors.ErrTrackNotFound(trackID)
 	}
 
 	return nil
 }
 
-func (s *SDKSource) subscribeToExpectedTracks(expecting map[string]bool) error {
-	// copy the array as the participants list may change as we walk it
-	participants := s.room.GetParticipants()
-
-	for _, p := range participants {
-		for _, track := range p.Tracks() {
-			if expecting[track.SID()] {
-				if rt, ok := track.(*lksdk.RemoteTrackPublication); ok {
-					err := rt.SetSubscribed(true)
-					if err != nil {
-						return err
-					}
-
-					delete(expecting, track.SID())
-					s.active.Inc()
-					if len(expecting) == 0 {
-						return nil
-					}
-				}
-			}
-		}
+func (s *SDKInput) onTrackPublished(pub *lksdk.RemoteTrackPublication, rp *lksdk.RemoteParticipant) {
+	if rp.Identity() != s.participantIdentity {
+		return
 	}
 
-	return nil
+	s.logger.Errorw("participant published new track", nil, "identity", s.participantIdentity)
 }
 
-func (s *SDKSource) onTrackMuted(pub lksdk.TrackPublication, _ lksdk.Participant) {
+func (s *SDKInput) onTrackMuted(pub lksdk.TrackPublication, _ lksdk.Participant) {
 	track := pub.Track()
 	if track == nil {
 		return
@@ -283,7 +188,7 @@ func (s *SDKSource) onTrackMuted(pub lksdk.TrackPublication, _ lksdk.Participant
 	}
 }
 
-func (s *SDKSource) onTrackUnmuted(pub lksdk.TrackPublication, _ lksdk.Participant) {
+func (s *SDKInput) onTrackUnmuted(pub lksdk.TrackPublication, _ lksdk.Participant) {
 	track := pub.Track()
 	if track == nil {
 		return
@@ -299,17 +204,17 @@ func (s *SDKSource) onTrackUnmuted(pub lksdk.TrackPublication, _ lksdk.Participa
 	}
 }
 
-func (s *SDKSource) onTrackUnpublished(track *lksdk.RemoteTrackPublication, _ *lksdk.RemoteParticipant) {
+func (s *SDKInput) onTrackUnpublished(track *lksdk.RemoteTrackPublication, _ *lksdk.RemoteParticipant) {
 	if w := s.getWriterForTrack(track.SID()); w != nil {
 		w.sendEOS()
 	}
 
 	if s.active.Dec() == 0 {
-		s.onComplete()
+		s.onDisconnected()
 	}
 }
 
-func (s *SDKSource) onComplete() {
+func (s *SDKInput) onDisconnected() {
 	select {
 	case <-s.endRecording:
 		return
@@ -318,7 +223,62 @@ func (s *SDKSource) onComplete() {
 	}
 }
 
-func (s *SDKSource) getWriterForTrack(trackID string) *appWriter {
+func (s *SDKInput) subscribeToParticipant(wg *sync.WaitGroup) error {
+	deadline := time.Now().Add(subscriptionTimeout)
+	for time.Now().Before(deadline) {
+		for _, p := range s.room.GetParticipants() {
+			if p.Identity() == s.participantIdentity {
+				for _, track := range p.Tracks() {
+					wg.Add(1)
+					if rt, ok := track.(*lksdk.RemoteTrackPublication); ok {
+						err := rt.SetSubscribed(true)
+						if err != nil {
+							return err
+						}
+					}
+				}
+
+				return nil
+			}
+		}
+		time.Sleep(100 * time.Millisecond)
+	}
+
+	return errors.ErrParticipantNotFound(s.participantIdentity)
+}
+
+func (s *SDKInput) subscribeToTracks(expecting map[string]struct{}) error {
+	deadline := time.Now().Add(subscriptionTimeout)
+	for time.Now().Before(deadline) {
+		for _, p := range s.room.GetParticipants() {
+			for _, track := range p.Tracks() {
+				if _, ok := expecting[track.SID()]; ok {
+					if rt, ok := track.(*lksdk.RemoteTrackPublication); ok {
+						err := rt.SetSubscribed(true)
+						if err != nil {
+							return err
+						}
+
+						delete(expecting, track.SID())
+						s.active.Inc()
+						if len(expecting) == 0 {
+							return nil
+						}
+					}
+				}
+			}
+		}
+		time.Sleep(100 * time.Millisecond)
+	}
+
+	for trackID := range expecting {
+		return errors.ErrTrackNotFound(trackID)
+	}
+
+	return nil
+}
+
+func (s *SDKInput) getWriterForTrack(trackID string) *appWriter {
 	switch trackID {
 	case s.trackID:
 		if s.audioWriter != nil {
@@ -333,72 +293,4 @@ func (s *SDKSource) getWriterForTrack(trackID string) *appWriter {
 	}
 
 	return nil
-}
-
-func (s *SDKSource) StartRecording() chan struct{} {
-	return nil
-}
-
-func (s *SDKSource) GetAudioSource() (*app.Source, webrtc.RTPCodecParameters) {
-	return s.audioSrc, s.audioCodec
-}
-
-func (s *SDKSource) GetVideoSource() (*app.Source, webrtc.RTPCodecParameters) {
-	return s.videoSrc, s.videoCodec
-}
-
-func (s *SDKSource) GetStartTime() int64 {
-	return s.cs.startTime.Load()
-}
-
-func (s *SDKSource) GetEndTime() int64 {
-	return s.cs.endTime.Load() + s.cs.delay.Load()
-}
-
-func (s *SDKSource) Playing(name string) {
-	var playing chan struct{}
-
-	if name == AudioAppSource {
-		playing = s.audioPlaying
-	} else if name == VideoAppSource {
-		playing = s.videoPlaying
-	} else {
-		return
-	}
-
-	select {
-	case <-playing:
-		return
-	default:
-		close(playing)
-	}
-}
-
-func (s *SDKSource) EndRecording() chan struct{} {
-	return s.endRecording
-}
-
-func (s *SDKSource) SendEOS() {
-	s.cs.SetEndTime(time.Now().UnixNano())
-
-	var wg sync.WaitGroup
-	if s.audioWriter != nil {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			s.audioWriter.sendEOS()
-		}()
-	}
-	if s.videoWriter != nil {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			s.videoWriter.sendEOS()
-		}()
-	}
-	wg.Wait()
-}
-
-func (s *SDKSource) Close() {
-	s.room.Disconnect()
 }
