+++ pkg/pipeline/output/websocket.go
@@ -6,11 +6,73 @@ import (
 	"net/http"
 
 	"github.com/gorilla/websocket"
+	"github.com/tinyzimmer/go-gst/gst"
+	"github.com/tinyzimmer/go-gst/gst/app"
 
 	"github.com/livekit/egress/pkg/errors"
 	"github.com/livekit/egress/pkg/pipeline/params"
 	"github.com/livekit/protocol/logger"
 )
+
+func buildWebsocketOutputBin(p *params.Params) (*OutputBin, error) {
+	writer, err := newWebSocketSink(p.WebsocketUrl, params.MimeTypeRaw, p.Logger, p.MutedChan)
+	if err != nil {
+		return nil, err
+	}
+
+	sink, err := app.NewAppSink()
+	if err != nil {
+		return nil, err
+	}
+
+	sink.SetCallbacks(&app.SinkCallbacks{
+		EOSFunc: func(appSink *app.Sink) {
+			// Close writer on EOS
+			if err = writer.Close(); err != nil && !errors.Is(err, io.EOF) {
+				p.Logger.Errorw("cannot close WS sink", err)
+			}
+		},
+		NewSampleFunc: func(appSink *app.Sink) gst.FlowReturn {
+			// Pull the sample that triggered this callback
+			sample := appSink.PullSample()
+			if sample == nil {
+				return gst.FlowEOS
+			}
+
+			// Retrieve the buffer from the sample
+			buffer := sample.GetBuffer()
+			if buffer == nil {
+				return gst.FlowError
+			}
+
+			// Map the buffer to READ operation
+			samples := buffer.Map(gst.MapRead).Bytes()
+
+			// From the extracted bytes, send to writer
+			_, err = writer.Write(samples)
+			if err != nil && !errors.Is(err, io.EOF) {
+				p.Logger.Errorw("cannot read AppSink samples", err)
+				return gst.FlowError
+			}
+			return gst.FlowOK
+		},
+	})
+
+	bin := gst.NewBin("output")
+	if err = bin.Add(sink.Element); err != nil {
+		return nil, err
+	}
+
+	ghostPad := gst.NewGhostPad("sink", sink.GetStaticPad("sink"))
+	if !bin.AddPad(ghostPad.Pad) {
+		return nil, errors.ErrGhostPadFailed
+	}
+
+	return &OutputBin{
+		bin:    bin,
+		logger: p.Logger,
+	}, nil
+}
 
 type websocketState string
