+++ pkg/pipeline/sink/playlist_writer.go
@@ -24,7 +24,6 @@ type PlaylistWriter struct {
 }
 
 func NewPlaylistWriter(p *params.Params) (*PlaylistWriter, error) {
-
 	// "github.com/grafov/m3u8" is fairly inefficient for frequent serializations of long playlists and
 	// doesn't implement recent additions to the HLS spec, but I'm not aware of anything better, short of
 	// writing one.
