--- pkg/pipeline/source/interface.go
@@ -1,7 +1,0 @@
-package source
-
-type Source interface {
-	StartRecording() chan struct{}
-	EndRecording() chan struct{}
-	Close()
-}
